+++ app/Http/Controllers/SizeController.php
@@ -78,6 +78,9 @@ class SizeController extends Controller
     public function destroy($id)
     {
         $size = Size::findOrFail($id);
+        if ($size->productVariants()->exists()) {
+            return redirect()->route('sizes.index')->with('error', 'Không thể xóa size vì đang được sử dụng trong sản phẩm.');
+        }
         $size->delete();
 
         return redirect()->route('sizes.index')->with('success', 'Xóa kích cỡ thành công!');
